refactor (package) : Model 데이터 내에 발생하는 exception 클래스 패키지 이동 (lotto.model -> lotto.model.exception)
