+++ src/main/java/study/lotto/model/exception/DuplicatedBonusLottoNumberException.java
@@ -1,4 +1,4 @@
-package study.lotto.model;
+package study.lotto.model.exception;
 
 public class DuplicatedBonusLottoNumberException extends IllegalArgumentException {
     public DuplicatedBonusLottoNumberException(final String message) {
