+++ src/main/java/study/lotto/model/exception/EmptyBonusLottoNumberException.java
@@ -1,4 +1,4 @@
-package study.lotto.model;
+package study.lotto.model.exception;
 
 public class EmptyBonusLottoNumberException extends IllegalArgumentException {
     public EmptyBonusLottoNumberException(final String message) {
