+++ src/main/java/study/lotto/model/exception/IllegalLotterySizeException.java
@@ -1,4 +1,4 @@
-package study.lotto.model;
+package study.lotto.model.exception;
 
 public class IllegalLotterySizeException extends IllegalArgumentException {
     public IllegalLotterySizeException(final String message) {
