+++ src/main/java/study/lotto/model/exception/MalFormedLottoNumberException.java
@@ -1,4 +1,4 @@
-package study.lotto.model;
+package study.lotto.model.exception;
 
 public class MalFormedLottoNumberException extends IllegalArgumentException {
     public MalFormedLottoNumberException(final String message) {
