+++ src/main/java/study/lotto/model/exception/NotEnoughMoneyException.java
@@ -1,4 +1,4 @@
-package study.lotto.model;
+package study.lotto.model.exception;
 
 public class NotEnoughMoneyException extends IllegalArgumentException {
     public NotEnoughMoneyException(final String message) {
