+++ src/main/java/study/lotto/model/exception/TicketLotteryBundleMustBeNotEmptyException.java
@@ -1,4 +1,4 @@
-package study.lotto.model;
+package study.lotto.model.exception;
 
 public class TicketLotteryBundleMustBeNotEmptyException extends IllegalArgumentException {
     public TicketLotteryBundleMustBeNotEmptyException(final String message) {
