+++ src/test/java/study/lotto/model/MoneyTest.java
@@ -2,6 +2,7 @@ package study.lotto.model;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import study.lotto.model.exception.NotEnoughMoneyException;
 
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
